WIP: Windows multiprocess wsgi
This change allows glance-api to use multiple workers on Windows
by not using fork, which is unavailable. Instead, we're duplicating
sockets and passing the handles through pipes. Also, instead of
using process groups, we have to go with Windows job objects.

Note that this doesn't change the workflow for other platforms.

Some tests will be added in a subsequent patchset.

Change-Id: Ic786199844e1d804962172286905036d93a4ed14